meshbook/requirements.txt
asyncio==3.4.3
configparser==7.1.0
pyyaml==6.0.2
websockets==14.1